Extract yasgui directive
## What
Wraps "ontotext-yasgui-web-component" with a directive.

## Why
Simplify integration of the component.

## How
Creates a directive and refactoring the "Sparql" view to use it.
