+++ src/js/angular/core/directives/yasgui-component/templates/yasgui-component.html
@@ -1,0 +1,16 @@
+<link href="css/lib/ontotext-yasgui-web-component.css?v=[AIV]{version}[/AIV]" rel="stylesheet">
+
+<ontotext-yasgui
+    ng-custom-element
+    ngce-prop-config="ontotextYasguiConfig"
+    ngce-prop-saved_query_config="savedQueryConfig"
+    ngce-prop-language="language"
+    ngce-on-create_saved_query="createSavedQuery($event)"
+    ngce-on-update_saved_query="updateSavedQuery($event)"
+    ngce-on-delete_saved_query="deleteSavedQuery($event)"
+    ngce-on-share_saved_query="shareSavedQuery($event)"
+    ngce-on-share_query="shareQuery($event)"
+    ngce-on-query_share_link_copied="queryShareLinkCopied()"
+    ngce-on-load_saved_queries="loadSavedQueries($event)"
+    ngce-on-output="output($event)">
+</ontotext-yasgui>
