+++ src/js/angular/sparql-template/controllers.js
@@ -4,10 +4,10 @@ import 'angular/rest/monitoring.rest.service';
 import 'angular/utils/notifications';
 import 'angular/utils/uri-utils';
 import 'services/ontotext-yasgui-web-component.service.js';
+import 'angular/core/services/event-emitter-service';
 import {decodeHTML} from "../../../app";
 import {DEFAULT_SPARQL_QUERY, SparqlTemplateInfo} from "../../../models/sparql-template/sparql-template-info";
 import {SparqlTemplateError} from "../../../models/sparql-template/sparql-template-error";
-import 'services/event-emitter-service';
 import {YasqeMode} from "../../../models/ontotext-yasgui/yasqe-mode";
 import {RenderingMode} from "../../../models/ontotext-yasgui/rendering-mode";
