+++ src/pages/sparql-editor.html
@@ -1,21 +1,7 @@
-<link href="css/lib/ontotext-yasgui-web-component.css?v=[AIV]{version}[/AIV]" rel="stylesheet">
 <div>
     <h1>YASGUI integration page</h1>
 
-    <ontotext-yasgui
-        ng-custom-element
-        ngce-prop-config="config"
-        ngce-prop-saved_query_config="savedQueryConfig"
-        ngce-prop-language="language"
-        ngce-on-create_saved_query="createSavedQuery($event)"
-        ngce-on-update_saved_query="updateSavedQuery($event)"
-        ngce-on-delete_saved_query="deleteSavedQuery($event)"
-        ngce-on-share_saved_query="shareSavedQuery($event)"
-        ngce-on-share_query="shareQuery($event)"
-        ngce-on-query_share_link_copied="queryShareLinkCopied()"
-        ngce-on-load_saved_queries="loadSavedQueries($event)"
-        ngce-on-output="output($event)">
-    </ontotext-yasgui>
+    <yasgui-component id="query-editor" yasgui-config="yasguiConfig" after-init="initViewFromUrlParams()"></yasgui-component>
 
     <form id="wb-download" method="GET" action="" target="_self">
         <input id="wb-download-query" name="query" type="hidden"/>
