+++ webpack.config.common.js
@@ -172,6 +172,11 @@ module.exports = {
                 transform: replaceVersion
             },
             {
+                from: 'src/js/angular/core/directives/yasgui-component/templates',
+                to: 'js/angular/core/directives/yasgui-component/templates',
+                transform: replaceVersion
+            },
+            {
                 from: 'src/js/angular/templates',
                 to: 'js/angular/templates'
             }
